requirements.txt
discord.py>=2.0.0
python-telegram-bot>=20.0
Pillow>=10.0.0
requests>=2.31.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
